+++ src/theme/components/link.ts
@@ -1,9 +1,23 @@
+import {
+  buttonBaseStyle,
+  callout,
+  noBrand,
+  pill,
+  primary,
+  secondary,
+} from "./button";
+
 export const baseLinkStyles = {
   color: "ui.link.primary",
   textDecoration: "underline",
   _hover: {
     color: "ui.link.secondary",
   },
+};
+
+const baseButtonLinkStyles = {
+  ...buttonBaseStyle,
+  display: "inline-flex",
 };
 
 const variants = {
@@ -26,6 +40,7 @@ const variants = {
       textDecoration: "underline",
     },
   },
+  // The "button" variant is deprecated.
   button: {
     width: "100px",
     borderRadius: "sm",
@@ -43,6 +58,43 @@ const variants = {
       color: "ui.white",
       bg: "ui.link.secondary",
       textDecoration: "none",
+    },
+  },
+  buttonPrimary: {
+    ...baseButtonLinkStyles,
+    ...primary({}),
+  },
+  buttonSecondary: {
+    ...baseButtonLinkStyles,
+    ...secondary({}),
+    _visited: {
+      color: "ui.link.primary",
+    },
+  },
+  buttonPill: {
+    ...baseButtonLinkStyles,
+    ...pill({}),
+    _visited: {
+      color: "ui.black",
+    },
+  },
+  buttonCallout: {
+    ...baseButtonLinkStyles,
+    ...callout({}),
+  },
+  buttonNoBrand: {
+    ...baseButtonLinkStyles,
+    ...noBrand({}),
+  },
+  buttonDisabled: {
+    ...baseButtonLinkStyles,
+    ...primary({}),
+    bg: "ui.gray.light-cool",
+    color: "ui.gray.dark",
+    opacity: "1",
+    pointerEvents: "none",
+    _visited: {
+      color: "ui.gray.dark",
     },
   },
 };